web-memory-timeline/tools/requirements_faiss.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
requests==2.32.3
